refactor: extract question display and options into separate components
