+++ types/algorithm-pattern.ts
@@ -1,0 +1,25 @@
+export type AlgorithmPattern = {
+	id: string;
+	name: string;
+};
+
+export const algorithmPatterns: AlgorithmPattern[] = [
+	{ id: "slidingWindow", name: "Sliding Window" },
+	{ id: "twoPointer", name: "Two Pointer" },
+	{ id: "fastSlowPointers", name: "Fast Slow Pointers" },
+	{ id: "binarySearch", name: "Binary Search" },
+	{ id: "heapTopK", name: "Heap Top K" },
+	{ id: "bfs", name: "BFS" },
+	{ id: "dfs", name: "DFS" },
+	{ id: "bitwise", name: "Bitwise" },
+	{ id: "backtracking", name: "Backtracking" },
+	{ id: "dynamicProgramming1d", name: "Dynamic Programming 1D" },
+	{ id: "dynamicProgramming2d", name: "Dynamic Programming 2D" },
+	{ id: "greedy", name: "Greedy" },
+	{ id: "stack", name: "Stack" },
+	{ id: "mergeIntervals", name: "Merge Intervals" },
+	{ id: "math", name: "Math" },
+	{ id: "trees", name: "Trees" },
+	{ id: "hashing", name: "Hashing" },
+	{ id: "linkedList", name: "Linked List" },
+];
